Add support for snippets that send output toPlaceholder
This does not work generically for getPage; the Snippet that is being called via getPage must implement the toPlaceholder property and send it's output to that placeholder. getPage will simply grab the output from the placeholder for caching purposes if specified, and can reset the output into the placeholder from cache.